@@ -1,34 +1,50 @@
 package domain;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
 public class LottoTicketStore {
-    public static int LOTTO_TICKET_PRICE = 1_000;
-    private final LottoTicketGenerator lottoTicketGenerator;
-
-    public LottoTicketStore(LottoTicketGenerator lottoTicketGenerator){
-        this.lottoTicketGenerator = lottoTicketGenerator;
-    }
-    public List<LottoTicket> purchaseLotto(List<List<LottoNumber>> lottoNumbers, Wallet wallet) {
-        int totalPrice = lottoNumbers.size() * LOTTO_TICKET_PRICE;
-
-        if(totalPrice > wallet.getRemainAmount()){
-            throw new IllegalArgumentException("구입금액이 부족합니다.");
+    public class LottoTicket{
+        private final LottoNumbers lottoNumbers;
+        private LottoTicket(LottoNumbers lottoNumbers){
+            this.lottoNumbers = lottoNumbers;
         }
 
-        wallet.use(totalPrice);
+        public LottoNumbers getLottoNumbers() {
+            return lottoNumbers;
+        }
+    }
+
+    public static final int MANUAL_LOTTO_TICKET_COST = 1_000;
+    public static final int AUTO_LOTTO_TICKET_COST = 1_000;
+
+    private final LottoNumbersGenerator lottoNumbersGenerator;
+
+    public LottoTicketStore(LottoNumbersGenerator lottoNumbersGenerator){
+        this.lottoNumbersGenerator = lottoNumbersGenerator;
+    }
+
+    public List<LottoTicket> purchaseManualLotto(List<LottoNumbers> lottoNumbers, int paymentCost) {
+        int totalCost = lottoNumbers.size() * MANUAL_LOTTO_TICKET_COST;
+
+        validationCost(paymentCost, totalCost);
 
         return lottoNumbers.stream()
                 .map(LottoTicket::new)
-                .collect(Collectors.toList());
+                .toList();
     }
 
-    public List<LottoTicket> purchaseLotto(Wallet wallet) {
-        int lottoCount = wallet.getRemainAmount() / LOTTO_TICKET_PRICE;
-        int totalPrice = lottoCount * LOTTO_TICKET_PRICE;
+    public List<LottoTicket> purchaseAutoLotto(int purchaseLottoCount, int paymentCost) {
+        int totalCost = purchaseLottoCount * AUTO_LOTTO_TICKET_COST;
 
-        wallet.use(totalPrice);
-        return lottoTicketGenerator.generate(lottoCount);
+        validationCost(paymentCost, totalCost);
+        return lottoNumbersGenerator.generate(purchaseLottoCount).stream()
+                .map(LottoTicket::new)
+                .toList();
+    }
+
+    private static void validationCost(int paymentCost, int totalCost) {
+        if(totalCost > paymentCost){
+            throw new IllegalArgumentException("구입금액이 부족합니다.");
+        }
     }
 }
