@@ -1,7 +1,6 @@
 package util;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
 public class IntegerUtil {
     public static boolean isInteger(String text) {
@@ -17,6 +16,6 @@ public class IntegerUtil {
         return numbers.stream()
                 .map(String::trim)
                 .map(Integer::parseInt)
-                .collect(Collectors.toList());
+                .toList();
     }
 }
