@@ -1,6 +1,5 @@
 import controller.LottoSimulator;
-import controller.QuickPickSimulator;
-import domain.LottoTicketAutoGenerator;
+import domain.RandomLottoNumbersGenerator;
 import domain.LottoTicketStore;
 import view.InputView;
 import view.OutputView;
@@ -9,9 +8,9 @@ public class Application {
     public static void main(String[] args) {
         InputView inputView = new InputView(System.out, System.in);
         OutputView outputView = new OutputView(System.out);
-        LottoTicketStore lottoTicketStore = new LottoTicketStore(new LottoTicketAutoGenerator());
+        LottoTicketStore lottoTicketStore = new LottoTicketStore(new RandomLottoNumbersGenerator());
 
-        LottoSimulator lottoSimulator = new QuickPickSimulator(inputView, outputView, lottoTicketStore);
+        LottoSimulator lottoSimulator = new LottoSimulator(inputView, outputView, lottoTicketStore);
         lottoSimulator.simulate();
     }
 }
