@@ -3,5 +3,5 @@ package domain;
 public class LottoConstant {
     public static int LOTTO_NUMBER_MIN_VALUE = 1;
     public static int LOTTO_NUMBER_MAX_VALUE = 45;
-    public static final int LOTTO_TICKET_LENGTH = 6;
+    public static final int LOTTO_NUMBERS_LENGTH = 6;
 }
