@@ -3,8 +3,8 @@ package domain;
 import java.util.Objects;
 
 public class LottoMatchResult {
-    public static final String INVALID_MATCH_COUNT_AND_IS_BONUS_NUMBER_MATCHED_MSG = "해당 일치 수와 보너스 일치 여부는 적절하지 않습니다.";
-    public static final int MATCH_COUNT_MIN_VALUE = 0;
+    private static final String INVALID_MATCH_COUNT_AND_IS_BONUS_NUMBER_MATCHED_MSG = "해당 일치 수와 보너스 일치 여부는 적절하지 않습니다.";
+    private static final int MATCH_COUNT_MIN_VALUE = 0;
     private final int matchCount;
     private final boolean isBonusNumberMatched;
 
@@ -17,7 +17,7 @@ public class LottoMatchResult {
     }
 
     private boolean isValidLottoMatchResult(int matchCount, boolean isBonusNumberMatched) {
-        if(matchCount == LottoConstant.LOTTO_TICKET_LENGTH && isBonusNumberMatched
+        if(matchCount == LottoConstant.LOTTO_NUMBERS_LENGTH && isBonusNumberMatched
             || matchCount < MATCH_COUNT_MIN_VALUE){
             return false;
         }
