@@ -5,7 +5,7 @@ import java.util.Objects;
 import static domain.LottoConstant.*;
 
 public class LottoNumber {
-    public static final String INVALID_LOTTO_NUMBER_MSG = String.format("로또 번호는 %d 이상 %d 이하여야합니다.", LOTTO_NUMBER_MIN_VALUE, LOTTO_NUMBER_MAX_VALUE);
+    private static final String INVALID_LOTTO_NUMBER_MSG = String.format("로또 번호는 %d 이상 %d 이하여야합니다.", LOTTO_NUMBER_MIN_VALUE, LOTTO_NUMBER_MAX_VALUE);
     private final int number;
 
     public LottoNumber(int number) {
