@@ -5,6 +5,7 @@ import domain.*;
 import java.io.PrintStream;
 import java.util.List;
 import java.util.stream.Collectors;
+import domain.LottoTicketStore.LottoTicket;
 
 import static domain.LottoRank.*;
 
@@ -28,7 +29,7 @@ public class OutputView {
 
     private void printLottoTickets(List<LottoTicket> lottoTickets) {
         for(LottoTicket lottoTicket : lottoTickets){
-            List<LottoNumber> lottoNumbers = lottoTicket.getLottoNumbers();
+            LottoNumbers lottoNumbers = lottoTicket.getLottoNumbers();
             outputStream.print("[");
             outputStream.print(joinLottoNumbers(lottoNumbers));
             outputStream.print("]");
@@ -36,8 +37,8 @@ public class OutputView {
         }
     }
 
-    private static String joinLottoNumbers(List<LottoNumber> lottoNumbers) {
-        return lottoNumbers.stream()
+    private static String joinLottoNumbers(LottoNumbers lottoNumbers) {
+        return lottoNumbers.getLottoNumbers().stream()
                 .map(lottoNumber -> Integer.toString(lottoNumber.getNumber()))
                 .collect(Collectors.joining(", "));
     }
